+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
 ---------------------
+##### Added
+- Duration field to module response.
+
 ##### Fixed
 - Path error when using library in an app such as enketo-validate-webapp.